pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "giji"
version = "1.1.0"
description = "Automation toolkit for importing GitHub issues into Jira"
authors = [
    { name = "Yustina Kvrivishvili", email = "tangelapo@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }

dependencies = [
    "requests>=2.31.0",
    "PyYAML>=6.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "flake8>=7.0.0",
    "pylint>=3.0.0",
    "mypy>=1.10.0",
    "tox>=4.14.0",
    "pytest>=8.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "scripts*"]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
explicit_package_bases = true
check_untyped_defs = true
disallow_untyped_defs = false
warn_unused_ignores = true
warn_return_any = false
pretty = true

[tool.pylint.'MASTER']
ignore-paths = ["build", "dist", "venv", "__pycache__"]
max-line-length = 120
disable = ["C0114", "C0115", "C0116"]

[tool.pylint.messages_control]
enable = ["W", "E"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
